Fix parsing of command argument
Apparently on older versions of Python set_defaults() does not overwrite
values from the main parser when called from a subparser. This breaks
the program completely for those version.

The fix is to not set the func parameter in the main parser and check
its existance later. This seems cleaner anyway.

Signed-off-by: Konrad Gräfe <konradgraefe@aol.com>